Separate milestone modals to their own components
